content/canvas/test/test_2d.pattern.basic.type.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.pattern.basic.type</title>
<script src="/MochiKit/MochiKit.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>
var _deferred = false;

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

var img = document.getElementById('green.png');
var pattern = ctx.createPattern(img, 'no-repeat');
ok(Object.prototype.toString.call(pattern) === '[object CanvasPattern]', "Object.prototype.toString.call(pattern) === '[object CanvasPattern]'");

if (!_deferred) SimpleTest.finish();
});
</script>
<img src="image_green.png" id="green.png" class="resource">
